requirements.txt
Django==4.2.7
pyecharts==2.0.4
mysqlclient==2.2.0
pandas==2.1.1
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0 